Disable bind mounts on Windows to allow pfe to start. Move docker-compose code to docker.go.
